Refactor common functionality into projectGenerator and make it configurable
