+++ packages/nx-plugin/src/generators/project/jest/index.ts
@@ -1,2 +1,3 @@
 export * from "./jestProjectGenerator";
 export { jestProjectGenerator as default } from "./jestProjectGenerator";
+export * from "./jestProjectGenerator.schema";
