+++ packages/nx-plugin/src/generators/project/jest/jestProjectGenerator.schema.ts
@@ -1,5 +1,9 @@
-import { ProjectGeneratorSchema } from "../projectGenerator.schema";
+import { ProjectGeneratorBaseSchema } from "../projectGeneratorBase.schema";
 
-export interface JestProjectGeneratorSchema extends ProjectGeneratorSchema {
+export interface JestProjectGeneratorSchema
+  extends JestProjectGeneratorOwnOptions,
+    ProjectGeneratorBaseSchema {}
+
+export interface JestProjectGeneratorOwnOptions {
   testEnvironment?: "node" | "jsdom";
 }
