+++ packages/nx-plugin/src/generators/project/node/nodeProjectGenerator.schema.ts
@@ -1,6 +1,6 @@
-import { ProjectGeneratorSchema } from "../projectGenerator.schema";
+import { ProjectGeneratorBaseSchema } from "../projectGeneratorBase.schema";
 
-export interface NodeProjectGeneratorSchema extends ProjectGeneratorSchema {
+export interface NodeProjectGeneratorSchema extends ProjectGeneratorBaseSchema {
   rootProjectGeneratorName: string;
   tags?: string;
 }
