+++ packages/nx-plugin/src/generators/project/project/index.ts
@@ -1,0 +1,2 @@
+export * from "./projectGenerator";
+export * from "./projectGenerator.schema";
