+++ packages/nx-plugin/src/generators/project/project/projectGenerator.schema.ts
@@ -1,0 +1,8 @@
+import { TsConfigProjectGeneratorOwnOptions } from "..";
+import { JestProjectGeneratorOwnOptions } from "../jest/index";
+import { ProjectGeneratorBaseSchema } from "../projectGeneratorBase.schema";
+
+export interface ProjectGeneratorSchema extends ProjectGeneratorBaseSchema {
+  jest?: JestProjectGeneratorOwnOptions;
+  tsconfig?: TsConfigProjectGeneratorOwnOptions;
+}
