+++ packages/nx-plugin/src/generators/project/tsconfig/index.ts
@@ -1,3 +1,4 @@
 export * from "./tsconfigProjectGenerator";
 export { tsconfigProjectGenerator as default } from "./tsconfigProjectGenerator";
 export * from "./tsconfigProjectGenerator.presets";
+export * from "./tsconfigProjectGenerator.schema";
