+++ packages/nx-plugin/src/generators/project/tsconfig/tsconfigProjectGenerator.presets.ts
@@ -1,4 +1,9 @@
 export const TsConfigGeneratorPresets = {
+  lib: {
+    lib: ["es2015"],
+    module: "commonjs",
+    target: "es2015",
+  },
   node18: {
     appLibTypes: ["node"],
     lib: ["es2022"],
